[close-failing-tests-stale-issues-task] Cleaned code, applied some suggestions
